requirements.txt
Django>=3.1.6
gunicorn>=20.0.4
dj-database-url>=0.5.0
whitenoise>=5.2
djangorestframework>=3.12
psycopg2>=2.8
requests>=2.25
